KCore: cleanup
git-svn-id: https://elsa-svn.onera.fr/CASSIOPEE/Trunk/Cassiopee/Apps/Modules@4438 2bbd7fe0-4929-0410-bb6b-e55522daae51

+++ KCore/KCore/Fld/FldFortranVecF.for
@@ -77,9 +77,9 @@ C =============================================================================
       IMPLICIT NONE
 C_IN
       INTEGER_E length
-      INTEGER rhs(length)
+      INTEGER*4 rhs(length)
 C_OUT
-      INTEGER lhs(length)
+      INTEGER*4 lhs(length)
 C_LOCAL
       INTEGER_E i
